Add main util cli
