+++ Cargo.toml
@@ -29,7 +29,7 @@ members = [
   "location",
   "mcu-interface",
   "mcu-util",
-  "orb-info",
+  "orb-info", 
   "qr-link",
   "s3-helpers",
   "security-utils",
@@ -48,6 +48,7 @@ members = [
   "update-agent",
   "update-agent/core",
   "update-verifier",
+  "util",
   "verity-tree-calc",
   "wpa-supplicant",
   "wt-video",
